Api Producto asincróna con los siguiente elementos:
Se creo el api de prodcuto con asincronismo, se crearon los metdos get,
 post, put, delete, con 5 packages(models, repository, services, controoler, confing)

Se implemento  la base datos H2, y el puerto de arranqué de la aplicación es 8089.

+++ src/main/java/com/legalSuite/Catalogo_productos/services/ProductoService.java
@@ -4,10 +4,12 @@ import com.legalSuite.Catalogo_productos.models.ProductoEntity;
 import com.legalSuite.Catalogo_productos.repository.IProductoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.data.domain.Pageable;
 import java.util.List;
 import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
 
 @Service
 public class ProductoService {
@@ -15,32 +17,38 @@ public class ProductoService {
     @Autowired
     private IProductoService productoRepo;
 
-
-    public ProductoEntity saveProducto(ProductoEntity producto) {
-        return productoRepo.save(producto);
+    @Async
+    public CompletableFuture<ProductoEntity> saveProducto(ProductoEntity producto) {
+        return CompletableFuture.supplyAsync(() -> productoRepo.save(producto));
     }
 
-
+    @Async
     public Optional<List<ProductoEntity>> getProcto(){
         List<ProductoEntity> product = productoRepo.findAll();
         return Optional.of(product);
     }
-
-    public Optional<ProductoEntity> getProductoById(Long id){
-        return productoRepo.findById(id);
+    @Async
+    public CompletableFuture<Optional<ProductoEntity>> getProductoById(Long id) {
+        return CompletableFuture.completedFuture(productoRepo.findById(id));
     }
 
-    public Page<ProductoEntity> getProductosFilter(String nombre, Double precioMin, Double precioMax, Pageable pageable) {
-        return productoRepo.findByFilters(nombre, precioMin, precioMax, pageable);
+    @Async
+    public CompletableFuture<Page<ProductoEntity>> getProductosFilter(String nombre, Double precioMin, Double precioMax, Pageable pageable) {
+        return CompletableFuture.supplyAsync(() -> productoRepo.findByFilters(nombre, precioMin, precioMax, pageable));
     }
 
-    public Boolean deleteById(Long id){
+    @Async
+    public CompletableFuture<Boolean> deleteById(Long id) {
         try {
-            productoRepo.deleteById(id);
-            return Boolean.TRUE;
-        }catch (Exception e){
-            return Boolean.FALSE;
+            Optional<ProductoEntity> producto = productoRepo.findById(id);
+            if (producto.isPresent()) {
+                productoRepo.deleteById(id);
+                return CompletableFuture.completedFuture(Boolean.TRUE);
+            } else {
+                return CompletableFuture.completedFuture(Boolean.FALSE);
+            }
+        } catch (Exception e) {
+            return CompletableFuture.failedFuture(e);
         }
-
     }
 }
